move benchResults to bench_test.go
